+++ lua/rustaceanvim/rust_analyzer.lua
@@ -24,9 +24,25 @@ M.buf_request = function(bufnr, method, params, handler)
   if bufnr == nil or bufnr == 0 then
     bufnr = vim.api.nvim_get_current_buf()
   end
+  local client_found = false
   for _, client in ipairs(M.get_active_rustaceanvim_clients()) do
     if client.supports_method(method, { bufnr = bufnr }) then
       client.request(method, params, handler, bufnr)
+      client_found = true
+    end
+  end
+  if not client_found then
+    local error_msg = 'No rust-analyzer client for ' .. method .. ' attched to buffer ' .. bufnr
+    if handler then
+      ---@type lsp.HandlerContext
+      local ctx = {
+        bufnr = bufnr,
+        client_id = -1,
+        method = method,
+      }
+      handler({ code = -1, message = error_msg }, nil, ctx)
+    else
+      vim.notify(error_msg, vim.log.levels.ERROR)
     end
   end
 end
@@ -34,10 +50,15 @@ end
 ---@param method string LSP method name
 ---@param params table|nil Parameters to send to the server
 M.notify = function(method, params)
+  local client_found = false
   for _, client in ipairs(M.get_active_rustaceanvim_clients()) do
     if client.supports_method(method) then
       client.notify(method, params)
+      client_found = true
     end
+  end
+  if not client_found then
+    vim.notify('No rust-analyzer client found for method: ' .. method, vim.log.levels.ERROR)
   end
 end